mec/intelligent-agent/requirements.txt
blinker==1.9.0
build==1.2.2.post1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mercantile==1.2.1
numpy
packaging==25.0
paho-mqtt==2.1.0
pillow==11.2.1
pycparser==2.22
pygame==2.6.1
pymunk==6.11.1
pyproject_hooks==1.2.0
pyquadkey2==0.3.2
requests==2.32.4
setuptools==80.9.0
urllib3==2.5.0
utm==0.8.1
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
websocket-client
